fix: refactor, memperbaiki typedoc, menambahkan jsdoc
