+++ src/config/cards.ts
@@ -162,6 +162,10 @@ enum givenCardCondition {
 }
 
 export class CardPicker {
+  /**
+   * Pick a random card or wild card in ratio 16:1
+   * @returns Defined random card or wild card in ratio 16:1
+   */
   static pickRandomCard(): allCard {
     const status: randomCardCondition = randomWithBias(
       [randomCardCondition.randomCard, randomCardCondition.wild],
@@ -189,12 +193,21 @@ export class CardPicker {
     }
   }
 
+  /**
+   * Pick any card that appropriate as normal card
+   * @returns Appropriate initial card
+   */
   static getInitialCard() {
     return appropriateInitialCards[
       Math.floor(random() * appropriateInitialCards.length)
     ];
   }
 
+  /**
+   * Card picker by given card
+   * @param card A valid UNO card
+   * @returns Biased card by given card, random card, or initial card in ratio 13:5:1
+   */
   static pickCardByGivenCard(card: allCard): allCard {
     const status: givenCardCondition = randomWithBias(
       [
@@ -301,3 +314,104 @@ export class CardPicker {
     }
   }
 }
+
+/**
+ * A function that used to defining the state for switch case comparer
+ * @param firstState Card state from the deck
+ * @param secState Card state from the user given card
+ * @returns An objects that indicate a certain valid condition
+ */
+export const getSwitchState = (
+  firstState: IGetCardState,
+  secState: IGetCardState
+) => {
+  /**
+   * If the color or number is the same, but it's not special or plus4 card
+   */
+  const FIRSTCARD_IS_COLOR_OR_NUMBER_IS_SAME =
+    (firstState?.color === secState?.color ||
+      firstState?.number === secState?.number) &&
+    secState.state !== EGetCardState.VALID_SPECIAL;
+
+  /**
+   * If the first card is the wild and the color of second card is the same
+   * or the first card is the plus4 and the color of second card is the same
+   */
+  const FIRSTCARD_IS_WILD_OR_WILD4_IS_SAME_SECOND_COLOR =
+    (firstState.state === EGetCardState.VALID_WILD &&
+      firstState.color === secState.color) ||
+    (firstState.state === EGetCardState.VALID_WILD_PLUS4 &&
+      firstState.color === secState.color);
+
+  /**
+   * If the second card is special card and the color
+   * of the second card is the same as the first card color
+   */
+  const SECONDCARD_IS_VALIDSPECIAL_AND_SAME_COLOR_AS_FIRSTCARD =
+    secState.state === EGetCardState.VALID_SPECIAL &&
+    secState.color === firstState.color;
+
+  /**
+   * If the first card is special card and the type of
+   * the second card is the same as the first card type
+   */
+  const FIRSTCARD_IS_NTYPE_AND_SECONDCARD_IS_NTYPE_TOO =
+    firstState.state === EGetCardState.VALID_SPECIAL &&
+    secState.state === EGetCardState.VALID_SPECIAL &&
+    firstState.type === secState.type;
+
+  /**
+   * If the second card is wild card or in the other word is color only
+   */
+  const SECONDCARD_IS_WILD = secState.state === EGetCardState.VALID_WILD;
+
+  /**
+   * If the second card is plus4 card
+   */
+  const SECONDCARD_IS_WILD4 = secState.state === EGetCardState.VALID_WILD_PLUS4;
+
+  return {
+    FIRSTCARD_IS_COLOR_OR_NUMBER_IS_SAME,
+    FIRSTCARD_IS_WILD_OR_WILD4_IS_SAME_SECOND_COLOR,
+    SECONDCARD_IS_VALIDSPECIAL_AND_SAME_COLOR_AS_FIRSTCARD,
+    FIRSTCARD_IS_NTYPE_AND_SECONDCARD_IS_NTYPE_TOO,
+    SECONDCARD_IS_WILD,
+    SECONDCARD_IS_WILD4,
+  };
+};
+
+/**
+ * Card comparer for the main logic of the bot
+ * @param firstCard The card from the deck
+ * @param secCard The user given card
+ * @returns String that indicate is valid or not
+ */
+export const compareTwoCard = (firstCard: allCard, secCard: allCard) => {
+  const firstState = CardPicker.getCardState(firstCard);
+  const secState = CardPicker.getCardState(secCard);
+
+  const switchState = getSwitchState(firstState, secState);
+
+  switch (true) {
+    /* eslint-disable no-fallthrough */
+
+    // Valid wilddraw4 from player
+    case switchState.SECONDCARD_IS_WILD4:
+      return "STACK_PLUS_4";
+
+    // Valid wild color only from player
+    case switchState.SECONDCARD_IS_WILD:
+      return "STACK_WILD";
+
+    case switchState.FIRSTCARD_IS_COLOR_OR_NUMBER_IS_SAME:
+    case switchState.FIRSTCARD_IS_WILD_OR_WILD4_IS_SAME_SECOND_COLOR:
+      return "STACK";
+
+    case switchState.FIRSTCARD_IS_NTYPE_AND_SECONDCARD_IS_NTYPE_TOO:
+    case switchState.SECONDCARD_IS_VALIDSPECIAL_AND_SAME_COLOR_AS_FIRSTCARD:
+      return `VALID_SPECIAL_${secState.type!.toUpperCase()}`;
+
+    default:
+      return "UNMATCH";
+  }
+};
